dev: eliminacion del torneo completada

+++ Negocio/TorneoService.cs
@@ -73,7 +73,7 @@ namespace Negocio
 
                 if (torneo == null) throw new Exception("El torneo no tiene datos para salvar");
 
-                if (torneo.Fecha.Date.AddDays(-1) < DateTime.Today.Date) throw new Exception("El torneo no se puede suspender porque el excede el dia permitido");
+                if (torneo.Fecha.Date.AddDays(-1) < DateTime.Today.Date) throw new Exception("El torneo no se puede suspender porque los equipos están jugando o el torneo ha finalizado");
 
                 torneo.HabilitacionInscripcion = false;
                 torneo.Suspendido = true;
@@ -100,7 +100,10 @@ namespace Negocio
                                                            .SingleOrDefaultAsync(t => t.Id == torneoId);
 
                 if (torneoAEliminar == null) throw new Exception("El torneo que quiere eliminar no existe");
-                 if (torneoAEliminar.Inscripciones.Count > 0) throw new Exception("No puede eliminar porque hay equipos inscriptos");
+                if (torneoAEliminar.Inscripciones.Count > 0) throw new Exception("No puede eliminar porque hay equipos inscriptos");
+                if (torneoAEliminar.HabilitacionInscripcion) throw new Exception("No puede eliminar si tiene inscripciones abiertas");
+                if (torneoAEliminar.Suspendido) throw new Exception("No puede eliminar el torneo porque está suspendido");
+
                 _db.Torneos.Remove(torneoAEliminar);
                 int eliminacionRealizada = await _db.SaveChangesAsync();
